Fix Android build and import structure, generate Android Gradle files
- Update compileSdk/targetSdk to 37 for permission_handler compatibility
- Disable Kotlin incremental compilation (Windows photo_manager workaround)
- Fix model imports in main.dart, screens, and widgets
- Generate complete Android Gradle project structure
- Create assets directory for pubspec

+++ lib/widgets/summary_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../main.dart';
+import '../models/photo_group.dart';
 import '../utils/byte_formatter.dart';
 import '../services/deletion_service.dart';
 
@@ -17,8 +17,8 @@ class SummaryBar extends StatelessWidget {
     int totalReclaimable = 0;
 
     for (final group in groups) {
-      totalSelected += group.selectedCount;
-      totalReclaimable += group.reclaimableBytes;
+      totalSelected += group.selectedCount.toInt();
+      totalReclaimable += group.reclaimableBytes.toInt();
     }
 
     return Container(
